+++ src/main/java/com/ajou_nice/with_pet/admin/model/dto/AddWithPetServiceRequest.java
@@ -1,4 +1,4 @@
-package com.ajou_nice.with_pet.admin.model.criticalservice;
+package com.ajou_nice.with_pet.admin.model.dto;
 
 
 import javax.persistence.Lob;
@@ -13,24 +13,26 @@ import lombok.ToString;
 @NoArgsConstructor
 @Getter
 @ToString
-public class AddCriticalServiceRequest {
+public class AddWithPetServiceRequest {
 
 	@NotNull
 	private String serviceName;
-
 	@NotNull
+	@Lob
 	private String serviceImg;
-
 	@NotNull
 	private String serviceIntro;
 
 	@Data
-	public static class CriticalServiceModifyRequest{
+	public static class WithPetServiceModifyRequest{
 		@NotNull
 		private Long serviceId;
+
 		private String serviceName;
+
 		@Lob
 		private String serviceImg;
+
 		private String serviceIntroduction;
 	}
 }
