+++ src/main/java/com/ajou_nice/with_pet/domain/entity/PetSitterWithPetService.java
@@ -9,6 +9,8 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
+
+import com.ajou_nice.with_pet.withpet_service.model.entity.WithPetService;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
