+++ src/main/java/com/ajou_nice/with_pet/withpet_service/repository/WithPetServiceRepository.java
@@ -1,6 +1,6 @@
-package com.ajou_nice.with_pet.repository;
+package com.ajou_nice.with_pet.withpet_service.repository;
 
-import com.ajou_nice.with_pet.domain.entity.WithPetService;
+import com.ajou_nice.with_pet.withpet_service.model.entity.WithPetService;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
